backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter~=22.1
djoser==2.2.2
Pillow==10.1.0
psycopg2-binary==2.9.3
